♻️ Fix all broken tests caused by the migration of the external service *ClickUp* checking point abstraction.
